+++ src/net/frame_syncer.rs
@@ -12,7 +12,7 @@ pub(crate) struct FSPeer {
     /// If address is None, the peer is the current device.
     pub addr: Option<Addr>,
     pub name: heapless::String<16>,
-    pub friend_id: u32,
+    pub friend_id: u16,
     pub states: RingBuf<FrameState>,
 }
